requirements.txt
appdirs==1.4.3
astroid==2.0
beem==0.20.9
certifi==2018.10.15
chardet==3.0.4
Click==7.0
ecdsa==0.13
Events==0.3
future==0.17.0
gspread==3.0.1
httplib2==0.11.3
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.15.3
oauth2client==4.1.2
pep8==1.7.1
prettytable==0.7.2
PTable==0.9.2
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycryptodomex==3.6.6
pylibscrypt==1.7.1
pylint==2.0.0
python-dateutil==2.7.4
pytz==2018.6
requests==2.20.0
rsa==3.4.2
scrypt==0.8.6
six==1.11.0
typed-ast==1.1.0
urllib3==1.24
watson-developer-cloud==2.2.6
websocket-client==0.53.0
wrapt==1.10.11
